dataEngineeringProject/etl/requirements.txt
pandas
openpyxl
ipaddress
psycopg2-binary